+++ example/lib/main.dart
@@ -22,7 +22,17 @@ class WebViewApp extends StatefulWidget {
 }
 
 class _WebViewAppState extends State<WebViewApp> {
-  final workspaceConfiguration = {
+  // final BlocklyOptions workspaceConfiguration = BlocklyOptions(
+  //   grid: const GridOptions(
+  //     spacing: 20,
+  //     length: 3,
+  //     colour: '#ccc',
+  //     snap: true,
+  //   ),
+  //   toolbox: ToolboxInfo.fromJson(initialToolboxJson),
+  // );
+
+  final BlocklyOptions workspaceConfiguration = BlocklyOptions.fromJson(const {
     'grid': {
       'spacing': 20,
       'length': 3,
@@ -30,7 +40,32 @@ class _WebViewAppState extends State<WebViewApp> {
       'snap': true,
     },
     'toolbox': initialToolboxJson,
-  };
+    // null safety example
+    'collapse': null,
+    'comments': null,
+    'css': null,
+    'disable': null,
+    'horizontalLayout': null,
+    'maxBlocks': null,
+    'maxInstances': null,
+    'media': null,
+    'modalInputs': null,
+    'move': null,
+    'oneBasedIndex': null,
+    'readOnly': null,
+    'renderer': null,
+    'rendererOverrides': null,
+    'rtl': null,
+    'scrollbars': null,
+    'sounds': null,
+    'theme': null,
+    'toolboxPosition': null,
+    'trashcan': null,
+    'maxTrashcanContents': null,
+    'plugins': null,
+    'zoom': null,
+    'parentWorkspace': null,
+  });
 
   void onInject(state) {
     debugPrint('onInject: ${state.xml}\n${jsonEncode(state.json)}');