+++ lib/src/types/blockly_state.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/widgets.dart';
 
-/// Blockly state
+/// BlocklyState
 @immutable
 class BlocklyState {
-  /// Create a default Blockly state
+  /// BlocklyState constructor
   const BlocklyState({
     this.xml = '<xml xmlns="https://developers.google.com/blockly/xml"></xml>',
     this.json = const {},
@@ -15,14 +15,28 @@ class BlocklyState {
   /// json state
   final Map<String, dynamic>? json;
 
+  /// isNotEmpty method
+  bool isNotEmpty() {
+    return xml != null || json != null;
+  }
+
+  /// isEmpty method
+  bool isEmpty() {
+    return !isNotEmpty();
+  }
+
   /// jsonEncode
-  Map<String, dynamic> toJson() => {
-        'xml': xml,
-        'json': json,
-      };
+  Map<String, dynamic>? toJson() {
+    return isNotEmpty()
+        ? {
+            'xml': xml,
+            'json': json,
+          }
+        : null;
+  }
 
   /// state from json
-  BlocklyState.fromJson(Map<String, dynamic> data)
-      : xml = data['xml'],
-        json = data['json'];
+  BlocklyState.fromJson(Map<String, dynamic>? data)
+      : xml = data?['xml'],
+        json = data?['json'];
 }