+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this project will be documented in this file.
 
+# 1.2.0 (2023-12-19)
+
+* add Blockly types
+* add nullToUndefined - null safety
+
 # 1.1.0 (2023-12-16)
 
 * add web BlocklyEditor